Added Checkstyle and Javadoc
