+++ gutta-apievolution-java-codegen-plugin/src/main/java/gutta/apievolution/javacodegen/JavaEnumMember.java
@@ -1,5 +1,10 @@
 package gutta.apievolution.javacodegen;
 
+/**
+ * Representation of a Java enumeration member for code generation.
+ *
+ * <p/> <b>Note:</b> This class must be public for Velocity code generation to work.
+ */
 public class JavaEnumMember {
 
     private final String name;
@@ -8,7 +13,12 @@ public class JavaEnumMember {
         this.name = name;
     }
 
+    /**
+     * Returns this enum member's name.
+     * @return see above
+     */
     public String getName() {
         return this.name;
     }
+
 }
