+++ gutta-apievolution-java-codegen-plugin/src/main/java/gutta/apievolution/javacodegen/JavaInterface.java
@@ -5,9 +5,14 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+/**
+ * Representation of a Java interface for code generation.
+ *
+ * <p/> <b>Note:</b> This class must be public for Velocity code generation to work.
+ */
 public class JavaInterface extends JavaUserDefinedType {
 
-    public final List<JavaField> fields = new ArrayList<>();
+    private final List<JavaProperty> properties = new ArrayList<>();
 
     private final Set<String> fieldNames = new HashSet<>();
 
@@ -15,14 +20,18 @@ public class JavaInterface extends JavaUserDefinedType {
         super(packageName, name);
     }
 
-    public List<JavaField> getFields() {
-        return this.fields;
+    /**
+     * Returns the properties for which accessors are required in this interface.
+     * @return see above
+     */
+    public List<JavaProperty> getProperties() {
+        return this.properties;
     }
 
-    void addField(JavaField field) {
-        if (!this.fieldNames.contains(field.name)) {
-            this.fields.add(field);
-            this.fieldNames.add(field.name);
+    void addField(JavaProperty field) {
+        if (!this.fieldNames.contains(field.getName())) {
+            this.properties.add(field);
+            this.fieldNames.add(field.getName());
         }
     }
